+++ src/pages/Dashboard/Admin/AdminProfileMenu.jsx
@@ -1,6 +1,7 @@
 import {
     PiPencilLineThin,
     profileFieldMap,
+    FormInputComponent,
 } from "../../../components/barrel_module/Barrel.jsx";
 
 function AdminProfileMenu( { user } ) {
@@ -52,20 +53,7 @@ function AdminProfileMenu( { user } ) {
                         encType="multipart/form-data">
                             
                         {fieldMap.map((field, index) => (
-                            <div key={index} className="flex flex-col gap-2">
-                                <label htmlFor={field.name} className="text-gray-700 font-medium">
-                                    {field.label}
-                                </label>
-                                <input
-                                    type={field.type}
-                                    name={field.name}
-                                    id={field.name}
-                                    placeholder={field.placeholder}
-                                    className={"border border-gray-300 rounded-md px-4 py-3 focus:outline-none focus:ring-2 focus:ring-blue-500 " + field.addClass}
-                                    defaultValue={field.defaultValue}
-                                    disabled={field.disabled}
-                                />
-                            </div>
+                            <FormInputComponent field={field} key={index} />
                         ))}
                         <div className="flex flex-row gap-5 justify-end">
                             <button type="button" className="text-gray-700 font-medium px-8 py-2 rounded-xl border border-gray-300 hover:bg-gray-100/80">